wiseturn/requirements.txt
backcall==0.1.0
cachetools==3.0.0
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.3.0
defusedxml==0.5.0
Django==1.11.17
django-allauth==0.38.0
django-cors-headers==2.4.0
django-debug-toolbar==1.11
django-extensions==2.1.4
django-filter==2.0.0
django-neomodel==0.0.4
django-rest-auth==0.9.3
django-storages==1.7.1
djangorestframework==3.9.0
docutils==0.14
drf-yasg==1.14.0
google-api-core==1.7.0
google-auth==1.6.2
google-cloud-core==0.29.1
google-cloud-storage==1.13.2
google-resumable-media==0.3.2
googleapis-common-protos==1.5.6
gunicorn==19.9.0
idna==2.7
inflection==0.3.1
ipython==7.2.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.13.1
Jinja2==2.10
lucene-querybuilder==0.2
Markdown==3.0.1
MarkupSafe==1.1.1
neo4j-driver==1.6.2
neomodel==3.2.9
neotime==1.0.0
oauthlib==2.1.0
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
prompt-toolkit==2.0.7
protobuf==3.6.1
psycopg2==2.7.6.1
ptyprocess==0.6.0
pyasn1==0.4.5
pyasn1-modules==0.2.3
pydotplus==2.0.2
Pygments==2.3.0
pyparsing==2.3.1
python-dateutil==2.7.5
python3-openid==3.1.0
pytz==2018.7
requests==2.20.1
requests-oauthlib==1.0.0
rsa==4.0
ruamel.yaml==0.15.89
six==1.11.0
sqlparse==0.2.4
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.24.1
wcwidth==0.1.7
Werkzeug==0.14.1
